pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-adk>=1.27.2",
    "google-auth>=2.49.1",
    "google-cloud-aiplatform>=1.141.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.9.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-sdk>=1.38.0",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
